# chapter-2-blog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go 语言编程之旅: 一起用 Go 做项目",
        "title": "博客系统",
        "termsOfService": "https://github.com/ChrisLeejing/go-programming-tour-book-exercise",
        "contact": {},
        "version": "v1.0"
    },
    "paths": {
        "/api/v1/articles": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article 文章"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CreateArticleRequest",
                        "name": "CreateArticleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validate.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/model.TagSwagger"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag 标签"
                ],
                "summary": "获取多个标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "标签名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "default": 1,
                        "description": "状态",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/model.TagSwagger"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag 标签"
                ],
                "summary": "新增标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "CreateTagRequest",
                        "name": "CreateTagRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validate.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/model.TagSwagger"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/tags/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag 标签"
                ],
                "summary": "获取标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag 标签"
                ],
                "summary": "更新标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "UpdateTagRequest",
                        "name": "UpdateTagRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validate.UpdateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TagSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tag 标签"
                ],
                "summary": "删除标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "标签 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/auth": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token 鉴权"
                ],
                "summary": "获取认证token",
                "parameters": [
                    {
                        "description": "GetAuthRequest",
                        "name": "GetAuthRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validate.GetAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/upload/file": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File 文件"
                ],
                "summary": "文件上传",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件类型",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errcode.Error": {
            "type": "object"
        },
        "model.TagSwagger": {
            "type": "object"
        },
        "validate.CreateArticleRequest": {
            "type": "object",
            "required": [
                "created_by",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "cover_img_url": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "validate.CreateTagRequest": {
            "type": "object",
            "required": [
                "created_by"
            ],
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
        "validate.GetAuthRequest": {
            "type": "object",
            "required": [
                "app_key",
                "app_secret"
            ],
            "properties": {
                "app_key": {
                    "description": "https://github.com/gin-gonic/gin#model-binding-and-validation",
                    "type": "string"
                },
                "app_secret": {
                    "type": "string"
                }
            }
        },
        "validate.UpdateTagRequest": {
            "type": "object",
            "required": [
                "id",
                "modified_by"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "modified_by": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        }
    }
}